+++ src/r1/scr11a.c
@@ -225,7 +225,7 @@ scrflaga.w = scrflagb.w = scrflagc.w = scrflagz.w = 0;
   }
 
   scrollz_h((scra_hz << 5) + HspdWk, 64);
-  scrollc_h((long int)(scra_hz << 4) * 3 + (long int)HspdWk * 2, 16);
+  scrollc_h((scra_hz << 4) * 3 + HspdWk * 2, 16);
   scrollb_h((scra_hz << 7) + HspdWk, 4);
 
   yWk = scra_v_posit.w.h;
@@ -711,7 +711,7 @@ void scrollb_hv(int_union lD4, int_union lD5) {
 
     scrb_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= 4;
     }
     else {
@@ -728,7 +728,7 @@ lD0.l = lD3.l = scrb_v_posit.l;
 
     scrb_v_count ^= 16;
     lD0.l -= lD3.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= 1;
     }
     else {
@@ -775,7 +775,7 @@ void scrollb_h(Sint32 lD4, Sint32 flagb) {
   if (!lD1.b.b4) {
     scrb_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= flagb;
     }
     else {
@@ -803,7 +803,7 @@ void scrollc_h(Sint32 lD4, Sint32 flagc) {
   if (!lD1.b.b4) {
     scrc_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagc.b.h |= flagc;
     }
     else {
@@ -831,7 +831,7 @@ void scrollz_h(Sint32 lD4, Sint32 flagz) {
   if (!lD1.b.b4) {
     scrz_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagz.b.h |= flagz;
     }
     else {
