+++ src/r4/scr41b.c
@@ -204,7 +204,7 @@ scrflaga.w = scrflagb.w = scrflagc.w = scrflagz.w = 0;
   vscroll.w.h = scra_v_posit.w.h;
   vscroll.w.l = scrb_v_posit.w.h;
 
-temp1 = (long int)(scra_hz << 2) * 3;
+temp1 = (scra_hz << 2) * 3;
   temp2 = scra_vz << 7;
   scrollb_hv(temp1, temp2);
 
@@ -585,7 +585,7 @@ void scrollb_hv(Sint32 lD4, Sint32 lD5) {
 
     scrb_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= 4;
     }
     else {
@@ -596,14 +596,14 @@ void scrollb_hv(Sint32 lD4, Sint32 lD5) {
 
 lD0.l = lD3.l = scrb_v_posit.l;
   lD0.l += lD5;
-  if ((long int)lD0.l < 0) lD0.l = 0;
+  if (lD0.l < 0) lD0.l = 0;
   scrb_v_posit.l = lD0.l;
 
   if (((Uint8)(scrb_v_posit.w.h & 16) ^ scrb_v_count) == 0) {
 
     scrb_v_count ^= 16;
     lD0.l -= lD3.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= 1;
     }
     else {
@@ -650,7 +650,7 @@ void scrollb_h(Sint32 lD4, Sint32 flagb) {
   if (!lD1.b.b4) {
     scrb_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagb.b.h |= flagb;
     }
     else {
@@ -678,7 +678,7 @@ void scrollc_h(Sint32 lD4, Sint32 flagc) {
   if (!lD1.b.b4) {
     scrc_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagc.b.h |= flagc;
     }
     else {
@@ -706,7 +706,7 @@ void scrollz_h(Sint32 lD4, Sint32 flagz) {
   if (!lD1.b.b4) {
     scrz_h_count ^= 16;
     lD0.l -= lD2.l;
-    if ((long int)lD0.l < 0) {
+    if (lD0.l < 0) {
       scrflagz.b.h |= flagz;
     }
     else {
